backend/food_requirements.txt
# Food Health Rater API Requirements

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and Image Processing
google-generativeai==0.3.2
Pillow==10.1.0

# Data and Validation
pydantic==2.5.0
python-dotenv==1.0.0

# Optional for enhanced features
requests==2.31.0
numpy==1.25.2